Fixed libfreenect.h not found error when compiling the rtabmap-gui lib
git-svn-id: http://rtabmap.googlecode.com/svn/trunk/rtabmap@1313 f169173b-cf89-36c8-b27e-44dbe73f0c83

+++ guilib/src/MapVisibilityWidget.cpp
@@ -42,7 +42,7 @@ void MapVisibilityWidget::updateCheckBoxes()
 	QWidget * area = this->findChild<QWidget*>("area");
 	QVBoxLayout * layout = (QVBoxLayout *)area->layout();
 	QList<QCheckBox*> checkboxes = area->findChildren<QCheckBox*>();
-	while(checkboxes.size() && checkboxes.size() > _poses.size())
+	while(checkboxes.size() && checkboxes.size() > (int)_poses.size())
 	{
 		delete *checkboxes.begin();
 		checkboxes.erase(checkboxes.begin());
